requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-vite==3.0.1
djangorestframework==3.14.0
idna==3.6
inertia-django==0.5.3
mysqlclient==2.1.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
